django/requirements.txt
Django
django-environ
djangorestframework
psycopg2-binary
django-bootstrap4
gunicorn
django-simple-history
# Just useful for debugging
ipython
